Implement basic order functionality
It's a bit difficult to fully test this. We have to be careful to not
generate expenses from Digicert, and I have not been able to find a
sandbox to create dummy certificates in. That said, I believe this to be
mostly working for a very basic use case.

This change also includes a bit of refactoring in an effort to DRY up
some of code.