Change Organization ID expected type back to int
The Digicert API inexplicably began returning the Organization ID
attribute as a string earlier this week, and it has switched back to an
int. Considering it was the only ID attribute being returned as string,
I believe this is the correct, intended state.

A future todo is to explore the json.Number type and see if that helps
us be a bit more flexible and resilient to quirks like this.